Conflicts:
	platform/core-impl/src/com/intellij/psi/stubs/StubTree.java
	platform/platform-impl/src/com/intellij/openapi/vfs/impl/jar/JarHandler.java
	platform/platform-tests/testSrc/com/intellij/core/CoreModuleManagerTest.java

+++ plugins/groovy/src/org/jetbrains/plugins/groovy/refactoring/introduce/parameter/java2groovy/GroovyIntroduceParameterMethodUsagesProcessor.java
@@ -103,7 +103,7 @@ public class GroovyIntroduceParameterMethodUsagesProcessor implements IntroduceP
 
     final GrClosureSignatureUtil.ArgInfo<PsiElement>[] actualArgs = GrClosureSignatureUtil
       .mapParametersToArguments(signature, callExpression.getNamedArguments(), callExpression.getExpressionArguments(),
-                                callExpression.getClosureArguments(), callExpression, true, true);
+                                callExpression.getClosureArguments(), callExpression, false, true);
 
     final GroovyPsiElementFactory factory = GroovyPsiElementFactory.getInstance(data.getProject());
 
@@ -169,6 +169,7 @@ public class GroovyIntroduceParameterMethodUsagesProcessor implements IntroduceP
       public boolean execute(final int paramNum) {
         try {
           final GrClosureSignatureUtil.ArgInfo<PsiElement> actualArg = actualArgs[paramNum];
+          if (actualArg == null) return true;
           for (PsiElement arg : actualArg.args) {
             arg.delete();
           }
